Added auth functionality

+++ src/recipes/Controller/RecipeController.java
@@ -3,16 +3,21 @@ package recipes.Controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 import recipes.ExceptionHandler.BadRequestException;
+import recipes.ExceptionHandler.ForbiddenException;
 import recipes.ExceptionHandler.NotFoundException;
 import recipes.Model.Recipe;
 import recipes.Repository.RecipeRepository;
+import recipes.Repository.UserRepository;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 @RestController
 public class RecipeController {
@@ -20,6 +25,8 @@ public class RecipeController {
 
     @Autowired
     RecipeRepository recipeRepository;
+    @Autowired
+    private UserRepository userRepository;
 
     @GetMapping("/api/recipe/{id}")
     public ResponseEntity<Recipe> getRecipe(@PathVariable Long id) {
@@ -28,21 +35,24 @@ public class RecipeController {
     }
 
     @DeleteMapping("/api/recipe/{id}")
-    public ResponseEntity<String> deleteRecipe(@PathVariable Long id) {
+    public ResponseEntity<String> deleteRecipe(@AuthenticationPrincipal UserDetails userDetails, @PathVariable Long id) {
         Recipe recipe = recipeRepository.findRecipeById(id).orElseThrow(() -> new NotFoundException("Recipe not found!"));
+        if (!(Objects.equals(userRepository.findByEmailIgnoreCase(userDetails.getUsername()).orElseThrow(() -> new NotFoundException("User not found!")).getId(), recipe.getAuthorId()))) throw new ForbiddenException("Cannot delete non-owned recipes!");
         deleteRecipeFromRepo(recipe);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @PostMapping("/api/recipe/new")
-    public ResponseEntity<Map<String, Long>> saveRecipe(@RequestBody @Valid Recipe recipe) {
+    public ResponseEntity<Map<String, Long>> saveRecipe(@AuthenticationPrincipal UserDetails userDetails, @RequestBody @Valid Recipe recipe) {
+        recipe.setAuthorId(userRepository.findByEmailIgnoreCase(userDetails.getUsername()).orElseThrow(() -> new NotFoundException("User not found!")).getId());
         recipeRepository.save(recipe);
         return new ResponseEntity<>(Map.of("id", recipe.getId()), HttpStatus.OK);
     }
 
     @PutMapping("/api/recipe/{id}")
-    public ResponseEntity<String> updateRecipe(@PathVariable Long id, @RequestBody @Valid Recipe recipe){
+    public ResponseEntity<String> updateRecipe(@AuthenticationPrincipal UserDetails userDetails, @PathVariable Long id, @RequestBody @Valid Recipe recipe){
         Recipe toUpdate = recipeRepository.findRecipeById(id).orElseThrow(() -> new NotFoundException("Recipe not found!"));
+        if (!(Objects.equals(userRepository.findByEmailIgnoreCase(userDetails.getUsername()).orElseThrow(() -> new NotFoundException("User not found!")).getId(), toUpdate.getAuthorId()))) throw new ForbiddenException("Cannot delete non-owned recipes!");
         toUpdate.setName(recipe.getName());
         toUpdate.setCategory(recipe.getCategory());
         toUpdate.setDescription(recipe.getDescription());
